backend/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
alembic==1.10.2
    # via -r requirements.in
anyio==3.6.2
    # via
    #   httpcore
    #   starlette
    #   watchfiles
asyncpg==0.27.0
    # via -r requirements.in
certifi==2022.12.7
    # via
    #   httpcore
    #   httpx
cffi==1.15.1
    # via cryptography
click==8.1.3
    # via uvicorn
colorama==0.4.6
    # via
    #   click
    #   uvicorn
cryptography==40.0.1
    # via -r requirements.in
decorator==5.1.1
    # via validators
fastapi==0.95.0
    # via -r requirements.in
greenlet==2.0.2
    # via sqlalchemy
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==0.16.3
    # via httpx
httptools==0.5.0
    # via uvicorn
httpx==0.23.3
    # via -r requirements.in
idna==3.4
    # via
    #   anyio
    #   rfc3986
mako==1.2.4
    # via alembic
markupsafe==2.1.2
    # via mako
psycopg2==2.9.5
    # via -r requirements.in
pycparser==2.21
    # via cffi
pydantic==1.10.7
    # via
    #   fastapi
    #   sqlmodel
pyjwt==2.6.0
    # via -r requirements.in
python-dotenv==1.0.0
    # via
    #   -r requirements.in
    #   uvicorn
pyyaml==6.0
    # via uvicorn
rfc3986[idna2008]==1.5.0
    # via httpx
sniffio==1.3.0
    # via
    #   anyio
    #   httpcore
    #   httpx
sqlalchemy==1.4.41
    # via
    #   alembic
    #   sqlalchemy-utils
    #   sqlmodel
sqlalchemy-utils==0.40.0
    # via -r requirements.in
sqlalchemy2-stubs==0.0.2a32
    # via sqlmodel
sqlmodel==0.0.8
    # via -r requirements.in
starlette==0.26.1
    # via fastapi
typing-extensions==4.5.0
    # via
    #   alembic
    #   pydantic
    #   sqlalchemy2-stubs
uvicorn[standard]==0.21.1
    # via -r requirements.in
validators==0.20.0
    # via -r requirements.in
watchfiles==0.18.1
    # via uvicorn
websockets==10.4
    # via uvicorn
